requirements.txt
streamlit
docx2pdf
pywin32 
PyMuPDF
pyPDF2
pdf2image
pillow 
fpdf 
openpyxl 
xlrd
fitz
pyodbc
socket
user_agents
matplotlib
plotly
altair
datetime
pdfplumber
pyhanko